Reorganized controllers, station and observers into a src directory
