--- controllers/__init__.py
@@ -1,4 +1,0 @@
-from controllers.basic_controller import *
-from controllers.command_sequence_controller import *
-from controllers.command_sequence import *
-from controllers.point_cloud_controller import *
