+++ src/kinova_drake/observers/camera_viewer.py
